requirements.txt
absl-py==1.4.0
accelerate==0.25.0
aiofiles==23.2.1
aiohttp==3.9.1
aioprometheus==23.3.0
aiosignal==1.3.1
alembic==1.4.3
altair==5.0.1
anyio==3.7.1
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
asyncssh==2.14.1
attrs==23.1.0
backoff==2.2.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.6
cloud-sql-python-connector==1.4.3
cloudpickle==2.2.1
cmake==3.26.4
colorlog==6.7.0
contourpy==1.1.0
cryptography==41.0.7
cycler==0.11.0
Cython==0.29.35
docker==6.1.3
docker-pycreds==0.4.0
einops==0.7.0
etils==1.5.2
exceptiongroup==1.1.3
fastapi==0.101.1
faster-fifo==1.4.5
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.42.1
frozenlist==1.4.0
fschat==0.2.31
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
gradio==3.41.1
gradio_client==0.5.0
grpc-google-iam-v1==0.12.7
grpcio==1.59.3
grpcio-status==1.59.3
gym==0.23.0
gym-notices==0.0.8
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httptools==0.6.1
httpx==0.24.1
huggingface-hub==0.16.4
humanize==4.9.0
idna==3.4
immutabledict==4.0.0
importlib-metadata==6.7.0
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
kubernetes==28.1.0
line-profiler==4.0.3
lit==16.0.6
llvmlite==0.40.1
loguru==0.7.0
Mako==1.3.0
Markdown==3.4.3
markdown-it-py==3.0.0
markdown2==2.4.10
matplotlib==3.7.2
mdurl==0.1.2
minio==7.1.15
msgpack==1.0.5
msgpack-numpy==0.4.8
multidict==6.0.4
networkx==3.1
nh3==0.2.14
ninja==1.11.1
nle==0.9.0
numba==0.57.1
numpy==1.24.3
nvidia-ml-py==11.525.150
nvitop==1.1.2
oauthlib==3.2.2
opencv-python==4.7.0.72
orjson==3.9.5
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
pathtools==0.1.2
peft==0.6.2
Pillow=9.4.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.23.3
psutil==5.9.5
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycparser==2.21
pydantic==1.10.13
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytz==2023.3
pyzmq==25.1.0
quantile-python==1.1
ray==2.6.3
redis==4.5.5
referencing==0.30.2
regex==2023.8.8
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.28.1
setproctitle==1.3.2
shapely==2.0.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==1.2.19
sqlparse==0.4.4
starlette==0.27.0
svgwrite==1.4.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.1
termcolor==2.3.0
threadpoolctl==3.1.0
tiktoken==0.4.0
timeout-decorator==0.5.0
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.65.0
triton==2.1.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.19.0
vllm==0.4.1
wandb==0.15.5
watchfiles==0.21.0
wavedrom==2.0.3.post3
wcwidth==0.2.6
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==2.3.6
xmanager==0.4.0
yarl==1.9.3
zipp==3.15.0
zmq==0.0.0
-e ./rl_baseline/sample-factory
